# specs/004-fullstack-web-app/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: Todo Full-Stack Web Application API
  description: RESTful JSON API for managing user tasks with JWT authentication
  version: 2.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.example.com/api/v1
    description: Production server

tags:
  - name: tasks
    description: Task management operations

security:
  - BearerAuth: []

paths:
  /tasks:
    get:
      summary: List all tasks for authenticated user
      description: |
        Returns all tasks belonging to the authenticated user, with optional filtering and sorting.
        **Data Isolation**: Only returns tasks where `task.user_id == JWT.user_id`.
      tags:
        - tasks
      parameters:
        - name: status
          in: query
          description: Filter by completion status
          required: false
          schema:
            type: string
            enum: [pending, completed]
        - name: priority
          in: query
          description: Filter by priority level
          required: false
          schema:
            type: string
            enum: [high, medium, low]
        - name: tags
          in: query
          description: Filter by tags (comma-separated)
          required: false
          schema:
            type: string
            example: "work,urgent"
        - name: sort
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
            enum: [created_at, due_date, priority, title]
            default: created_at
        - name: order
          in: query
          description: Sort direction
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Successfully retrieved task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
              examples:
                emptyList:
                  summary: No tasks
                  value: []
                withTasks:
                  summary: User with tasks
                  value:
                    - id: "550e8400-e29b-41d4-a716-446655440000"
                      user_id: "650e8400-e29b-41d4-a716-446655440000"
                      title: "Buy groceries"
                      description: "Milk, eggs, bread"
                      completed: false
                      priority: "high"
                      tags: ["errands", "home"]
                      due_date: "2026-01-10"
                      recurrence: null
                      created_at: "2026-01-05T10:00:00Z"
                      updated_at: "2026-01-05T10:00:00Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new task
      description: |
        Creates a new task for the authenticated user.
        **Automatic Assignment**: `user_id` is automatically set from JWT; client cannot specify it.
      tags:
        - tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
            examples:
              minimal:
                summary: Minimal task (title only)
                value:
                  title: "Buy groceries"
              full:
                summary: Full task with all fields
                value:
                  title: "Complete project report"
                  description: "Include charts and analysis"
                  priority: "high"
                  tags: ["work", "urgent"]
                  due_date: "2026-01-15"
                  recurrence: "weekly"
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                user_id: "650e8400-e29b-41d4-a716-446655440000"
                title: "Buy groceries"
                description: null
                completed: false
                priority: null
                tags: []
                due_date: null
                recurrence: null
                created_at: "2026-01-05T10:00:00Z"
                updated_at: "2026-01-05T10:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /tasks/{id}:
    get:
      summary: Get a single task by ID
      description: |
        Returns a single task if it belongs to the authenticated user.
        **Ownership Check**: Returns 403 if task exists but belongs to another user.
      tags:
        - tasks
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Successfully retrieved task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update an existing task
      description: |
        Updates a task if it belongs to the authenticated user.
        **Ownership Check**: Returns 403 if task exists but belongs to another user.
        **Automatic Timestamps**: `updated_at` is automatically set to current time.
      tags:
        - tasks
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
            example:
              title: "Buy groceries (updated)"
              description: "Milk, eggs, bread, cheese"
              priority: "medium"
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete a task
      description: |
        Permanently deletes a task if it belongs to the authenticated user.
        **Ownership Check**: Returns 403 if task exists but belongs to another user.
      tags:
        - tasks
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '204':
          description: Task deleted successfully (no content)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /tasks/{id}/complete:
    patch:
      summary: Toggle task completion status
      description: |
        Toggles the completion status of a task (true ↔ false).
        **Recurring Tasks**: If task has recurrence and is marked complete, creates a NEW task with next due date.
        **Ownership Check**: Returns 403 if task exists but belongs to another user.
      tags:
        - tasks
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task completion status toggled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                user_id: "650e8400-e29b-41d4-a716-446655440000"
                title: "Buy groceries"
                description: null
                completed: true
                priority: null
                tags: []
                due_date: null
                recurrence: null
                created_at: "2026-01-05T10:00:00Z"
                updated_at: "2026-01-05T10:15:00Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token issued by Better Auth on frontend.
        Include in Authorization header: `Authorization: Bearer <token>`

  parameters:
    TaskId:
      name: id
      in: path
      description: UUID of the task
      required: true
      schema:
        type: string
        format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"

  schemas:
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title (required, non-empty)
          example: "Buy groceries"
        description:
          type: string
          maxLength: 2000
          nullable: true
          description: Detailed task description (optional)
          example: "Milk, eggs, bread"
        priority:
          type: string
          enum: [high, medium, low]
          nullable: true
          description: Task priority level (optional)
          example: "high"
        tags:
          type: array
          items:
            type: string
          default: []
          description: Array of string tags for categorization (optional)
          example: ["errands", "home"]
        due_date:
          type: string
          format: date
          nullable: true
          description: Task deadline date in YYYY-MM-DD format (optional)
          example: "2026-01-10"
        recurrence:
          type: string
          enum: [daily, weekly, monthly]
          nullable: true
          description: Recurrence pattern (optional)
          example: "weekly"

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title (optional in update)
          example: "Buy groceries (updated)"
        description:
          type: string
          maxLength: 2000
          nullable: true
          description: Detailed task description (optional)
          example: "Milk, eggs, bread, cheese"
        priority:
          type: string
          enum: [high, medium, low]
          nullable: true
          description: Task priority level (optional)
          example: "medium"
        tags:
          type: array
          items:
            type: string
          nullable: true
          description: Array of string tags (optional)
          example: ["errands"]
        due_date:
          type: string
          format: date
          nullable: true
          description: Task deadline date (optional)
          example: "2026-01-12"
        recurrence:
          type: string
          enum: [daily, weekly, monthly]
          nullable: true
          description: Recurrence pattern (optional)
          example: null

    TaskResponse:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - tags
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique task identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        user_id:
          type: string
          format: uuid
          description: Owner's user ID (from JWT)
          example: "650e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          description: Task title
          example: "Buy groceries"
        description:
          type: string
          nullable: true
          description: Task description
          example: "Milk, eggs, bread"
        completed:
          type: boolean
          description: Completion status
          example: false
        priority:
          type: string
          enum: [high, medium, low]
          nullable: true
          description: Priority level
          example: "high"
        tags:
          type: array
          items:
            type: string
          description: Array of tags
          example: ["errands", "home"]
        due_date:
          type: string
          format: date
          nullable: true
          description: Deadline date
          example: "2026-01-10"
        recurrence:
          type: string
          enum: [daily, weekly, monthly]
          nullable: true
          description: Recurrence pattern
          example: null
        created_at:
          type: string
          format: date-time
          description: Creation timestamp (ISO 8601)
          example: "2026-01-05T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp (ISO 8601)
          example: "2026-01-05T10:00:00Z"

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: "Task not found"

  responses:
    BadRequest:
      description: Invalid request (validation error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            emptyTitle:
              summary: Empty title
              value:
                detail: "Title cannot be empty"
            titleTooLong:
              summary: Title exceeds maximum length
              value:
                detail: "Title must be at most 200 characters"

    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingToken:
              summary: No Authorization header
              value:
                detail: "Not authenticated"
            invalidToken:
              summary: Invalid JWT signature
              value:
                detail: "Could not validate credentials"

    Forbidden:
      description: Valid token but insufficient permissions (task belongs to another user)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "You do not have permission to access this task"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Task not found"

    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "An unexpected error occurred"